requirements.txt
fitz==0.0.1.dev2
gensim==4.3.3
hdbscan==0.8.40
langdetect==1.0.9
matplotlib==3.10.6
nltk==3.9.1
numpy==2.3.3
openai==1.108.1
pandas==2.3.2
Pillow==11.3.0
protobuf==6.32.1
pytesseract==0.3.13
python-dotenv==1.1.1
rapidfuzz==3.14.1
scikit_learn==1.7.2
seaborn==0.13.2
sentence_transformers==5.1.0
spacy==3.8.7
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
umap==0.1.1
